refactor host datatype
